change file structure for serde modules
