+++ luact/serde/protobuf.lua
@@ -1,0 +1,5 @@
+local common = require 'luact.serde.common'
+local _M = {}
+
+
+return _M
