+++ test/tools/cluster.lua
@@ -84,6 +84,7 @@ function _M.start_luact(n_core, arbiter, proc)
 				fn()
 			end, function (e)
 				logger.error('err', e, debug.traceback())
+				os.exit(-2)
 			end)
 			luact.stop()
 		end, function (e)
